Avoid problems in case multiple Accessors use the same monitored item.
When using more than one Accessor with a monitored item the inital value
is read from the first Accessor. At this point a lock is used to protect
the data of the first Accessor read by the second Accessor. Else it
could be that the data of the first Accessors is just updated (e.g.
memory is cleared or new data is written).